+++ src/Yasumi/Provider/Canada.php
@@ -60,6 +60,7 @@ class Canada extends AbstractProvider
         $this->calculateLabourDay();
         $this->calculateThanksgivingDay();
         $this->calculateRemembranceDay();
+        $this->calculateNationalDayForTruthAndReconciliation();
     }
 
     public function getSources(): array
@@ -168,7 +169,11 @@ class Canada extends AbstractProvider
     /**
      * Canada Day.
      *
-     * @see https://en.wikipedia.org/wiki/Canada_Day
+     * @see https://en.wikipedia.org/wiki/Canada_Day.
+     * @see Holidays Act, R.S.C., 1985, c. H-5, https://laws-lois.justice.gc.ca/eng/acts/h-5/page-1.html
+     *
+     * by statute, Canada Day is July 1 if that day is not Sunday, and July 2 if July 1 is a Sunday.
+
      *
      * @throws InvalidDateException
      * @throws \InvalidArgumentException
@@ -180,11 +185,14 @@ class Canada extends AbstractProvider
         if ($this->year < 1983) {
             return;
         }
-
+        $date = new DateTime($this->year.'-07-01', DateTimeZoneFactory::getDateTimeZone($this->timezone));
+        if (7 === (int) $date->format('N')) {
+            $date = new DateTime($this->year.'-07-02', DateTimeZoneFactory::getDateTimeZone($this->timezone));
+        }
         $this->addHoliday(new Holiday(
             'canadaDay',
             [],
-            new DateTime($this->year.'-07-01', DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $date,
             $this->locale
         ));
     }
@@ -257,6 +265,30 @@ class Canada extends AbstractProvider
             'labourDay',
             [],
             new DateTime("first monday of september $this->year", DateTimeZoneFactory::getDateTimeZone($this->timezone)),
+            $this->locale
+        ));
+    }
+
+    /**
+     * National Day For Truth And Reconciliation.
+     *
+     * @see https://parl.ca/Content/Bills/432/Government/C-5/C-5_4/C-5_4.PDF, S. C. 2021, C.11.
+     *
+     * @throws InvalidDateException
+     * @throws \InvalidArgumentException
+     * @throws UnknownLocaleException
+     * @throws \Exception
+     */
+    private function calculateNationalDayForTruthAndReconciliation(): void
+    {
+        if ($this->year < 2021) {
+            return;
+        }
+
+        $this->addHoliday(new Holiday(
+            'truthAndReconciliationDay',
+            [],
+            new DateTime("last day of september $this->year", DateTimeZoneFactory::getDateTimeZone($this->timezone)),
             $this->locale
         ));
     }